Split the code in modules
