+++ src/index.js
@@ -1,58 +1,15 @@
 import './style.css';
+import getdata from './modules/getdata.js';
+import likesforthissearch from './modules/likesforthissearch.js';
+import getLikes from './modules/getlikes.js';
+import giveLikes from './modules/givelikes.js';
 // import getLikes from "./modules/gettinglikes.js";
-const artistName = 'nothing but thieves';
-const token = 'BQDqmlYc4abtA9aVrK4Z5c1UQwIkmT7yE7zuqKdlxvL7LvcJ2lz_hnI0Wp8LT5NCQseIsGPy2i-np9DdSZPj1mbJVQMthmAD_96QZVwwQABuY8-C1kfY3yPlfWRRjL6dZ33mzQK1pvbzXcDgk-qbK0Aj5Vq8ElVz4sR8uTyQ4vBGRlfvYD7xhPukoXF1NJjvsF4HA3IGnquJV7GE9c7x90taupJ5EmEoYNYL3L80M--rLLMI46O4aSk2gRSA-46fZ5OTQLzkIfh-YJFX3P4T';
 const musiccontainer = document.querySelector('.musiccontainer');
 const datanames = [];
-let updatedLikes = [];
 
-const likesforthissearch = async (data, likessaved) => {
-  const resolvedlikes = await likessaved;
-  const itemMap = new Map();
-  resolvedlikes.forEach((item) => {
-    itemMap.set(item.item_id, item.likes);
-  });
-  updatedLikes = data.tracks.items.map((item) => ({
-    item_id: item.id,
-    likes: itemMap.get(item.id),
-  }));
-};
-
-const getLikes = async () => {
-  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/N1YfreMOcnHDHjcZrEgf/likes/', {
-    method: 'GET',
-    headers: {
-      'Content-type': 'application/json',
-    },
-  });
-  if (response.ok) {
-    const data = await response.json();
-    return data;
-  } return false;
-};
-
-const giveLikes = async (itemidd) => {
-  await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/vYlIBG65vRC15spE8sZd/likes/', {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: `${itemidd}`,
-    }),
-    headers: {
-      'Content-type': 'application/json',
-    },
-  });
-};
-
-const getArtist = async () => {
-  const response = await fetch(`https://api.spotify.com/v1/search?q=${artistName}&type=track`, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const data = await response.json();
-
-  await likesforthissearch(data, getLikes());
-
+const render = async () => {
+  const data = await getdata();
+  const updatedLikes = await likesforthissearch(data, getLikes());
   for (let i = 0; i < data.tracks.items.length; i += 1) {
     const text = document.createElement('div');
     text.classList.add('songelementontainer', 'dflex', 'flexcol');
@@ -72,4 +29,6 @@ const getArtist = async () => {
   }
 };
 
-getArtist();
+document.addEventListener('DOMContentLoaded', () => {
+  render();
+});
